tia: initial RSYNC inregration
Docs suggest that RSYNC resets the HSYNC counter to the rising edge of
H@1, and then does a full reset after the next full H@1-H@2 cycle.

This isn't 100% accurate - I think - but it results in the sprite
positions in Pitfall being correct.

It does result in an off-by-one situation with each scanline though,
which I also notice in Pac-Man aswell.